+++ pyproject.toml
@@ -1,0 +1,27 @@
+[tool.poetry]
+name = "peewee-async"
+version = '0.8.1'
+description = "Asynchronous interface for peewee ORM powered by asyncio."
+authors = ["Alexey Kinev <rudy@05bit.com>", "Gorshkov Nikolay(contributor) <nogamemorebrain@gmail.com>"]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.7"
+peewee = "^3.15.4"
+
+aiopg = { version = "^1.4.0", optional = true }
+aiomysql = { version = "^0.2.0", optional = true }
+cryptography = { version = "^41.0.3", optional = true }
+pytest = { version = "^7.4.1", optional = true }
+pytest-asyncio = { version = "^0.21.1", optional = true }
+importlib-metadata= { version = "^6.7.0", optional = true }
+
+
+[tool.poetry.extras]
+postgresql = ["aiopg"]
+mysql = ["aiomysql", "cryptography"]
+develop = ["aiopg", "aiomysql", "cryptography", "importlib-metadata", "pytest", "pytest-asyncio"]
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
