+++ tests/test_obsolete.py
@@ -471,108 +471,3 @@ class ManagerTestCase(BaseManagerTestCase):
             self.assertEqual((obj_uuid, obj_alpha), 
                              (comp.uuid, comp.alpha))
         self.run_with_managers(test)
-
-
-######################
-# Transactions tests #
-######################
-
-
-class FakeUpdateError(Exception):
-    """Fake error while updating database.
-    """
-    pass
-
-
-class ManagerTransactionsTestCase(BaseManagerTestCase):
-    # only = ['postgres', 'postgres-ext', 'postgres-pool', 'postgres-pool-ext']
-    only = None
-
-    def test_atomic_success(self):
-        """Successful update in transaction.
-        """
-        async def test(objects):
-            obj = await objects.create(TestModel, text='FOO')
-            obj_id = obj.id
-
-            async with objects.atomic():
-                obj.text = 'BAR'
-                await objects.update(obj)
-
-            res = await objects.get(TestModel, id=obj_id)
-            self.assertEqual(res.text, 'BAR')
-
-        self.run_with_managers(test)
-
-    def test_atomic_failed(self):
-        """Failed update in transaction.
-        """
-        async def test(objects):
-            obj = await objects.create(TestModel, text='FOO')
-            obj_id = obj.id
-
-            try:
-                async with objects.atomic():
-                    obj.text = 'BAR'
-                    await objects.update(obj)
-                    raise FakeUpdateError()
-            except FakeUpdateError as e:
-                error = True
-                res = await objects.get(TestModel, id=obj_id)
-
-            self.assertTrue(error)
-            self.assertEqual(res.text, 'FOO')
-
-        self.run_with_managers(test)
-
-    def test_several_transactions(self):
-        """Run several transactions in parallel tasks.
-        """
-        wait = lambda tasks: self.loop.run_until_complete(
-            asyncio.wait([
-                self.loop.create_task(t) for t in tasks
-            ]))
-
-        async def t1(objects):
-            async with objects.atomic():
-                self.assertEqual(objects.database.transaction_depth_async(), 1)
-                await asyncio.sleep(0.25)
-
-        async def t2(objects):
-            async with objects.atomic():
-                self.assertEqual(objects.database.transaction_depth_async(), 1)
-                await asyncio.sleep(0.0625)
-
-        async def t3(objects):
-            async with objects.atomic():
-                self.assertEqual(objects.database.transaction_depth_async(), 1)
-                await asyncio.sleep(0.125)
-
-        for _, objects in self.managers.items():
-            wait([
-                t1(objects),
-                t2(objects),
-                t3(objects),
-            ])
-
-            with self.manager(objects, allow_sync=True):
-                for model in reversed(self.models):
-                    model.delete().execute()
-
-            self.run_count += 1
-
-    def test_atomic_fail_with_disconnect(self):
-        """Database gone in transaction.
-        """
-        async def test(objects):
-            error = False
-            try:
-                async with objects.atomic():
-                    await objects.database.close_async()
-                    raise FakeUpdateError()
-            except FakeUpdateError:
-                error = True
-
-            self.assertTrue(error)
-
-        self.run_with_managers(test)
